addressed gh comments
