+++ pax/conf/experiment/hyper.yaml
@@ -22,17 +22,6 @@ eval_every: 5000 # timesteps
 # num_updates = num_episodes / eval_every
 # batch_size = num_envs * num_steps
 
-# DQN agent parameters
-dqn: 
-  batch_size: 256
-  discount: 0.99
-  learning_rate: 1e-2
-  epsilon: 0.5
-  replay_capacity: 100000
-  min_replay_size: 1000
-  sgd_period: 1 
-  target_update_period: 4 
-
 # PPO agent parameters
 ppo:
   num_minibatches: 4
